+++ backend/src/main/resources/db/migration/V202109181800__create_tables.sql
@@ -1,5 +1,5 @@
 create table day_of_week_time_blocked (
-    id bigint not null,
+    id SERIAL not null,
     tenant varchar(255) not null,
     day_of_week INTEGER not null,
     start_time TIME not null,
